Improve NPC pathfinding in dynamic environments
Allow NPCs to not find a path for 10 ticks. This is sufficient to make an NPC adjust for changed blocks, rather than having the pathfind fallback teleport that NPC to its target straight away. It should be noted that NPC pathfinding can still fail for reasonable paths, which is a downside of using Minecraft's pathfinding implementation.